reddit_analyzer/requirements.txt
Django==3.1.5
nltk==3.5
numpy==1.19.5
textblob==0.15.3
requests==2.25.1
matplotlib==3.3.4
prawcore==1.5.0
wordcloud==1.8.1
praw==7.1.0
misaka==2.1.1
Pillow==8.1.0
django-bootstrap3==14.2.0
django-crispy-forms==1.11.0
gunicorn==20.0.4
whitenoise==5.2.0
